requirements.txt
fastapi
uvicorn
pydantic
httpx
torch
python-multipart
langgraph
langchain-text-splitters
langchain-core
langchain-community
langchain-ollama
pinecone-client
python-dotenv
tiktoken
pandas
numpy
requests
tqdm
arxiv